requirements.txt
psycopg2==2.6.1
stravalib==0.5.0
facebook-sdk==2.0.0
requests==2.12.2
